settings + noexcept + ODR fix + papi_rate_tuple + kokkos
- removed some extranous interface libraries
- reworked tim::settings to be more dynamic
- added noexcept specifiers in many places
- fixed ODR issue with manager
- using cuda implies using nvtx
- rusage metrics value reports last diff
- child timing functions
- miscellaneous optimizations
- auto timer is simpler
- ert and storage are now a global lib
- support for embedding interpreter
- re-worked kokkos + more kokkos configs
- timem supports intermediate output
- TIMEMORY_DISABLE_CUDA_HALF -> TIMEMORY_USE_CUDA_HALF

+++ source/tests/stl_overload_tests.cpp
@@ -79,7 +79,7 @@ class stl_overload_tests : public ::testing::Test
 #define GET(VAR, N) std::get<N>(VAR)
 
 constexpr size_t N         = 2;
-const double     tolerance = std::numeric_limits<float>::epsilon();
+const double     tolerance = static_cast<double>(std::numeric_limits<float>::epsilon());
 
 using int_type    = int;
 using real_type   = double;
